Add files via upload
-How do I interpret user needs and implement them into a program? How does creating “user stories” help with this?
Getting feedback is one of the most important aspects of development. Meeting with users and asking for input or simply having a bug report tool in the program can save countless hours of searching for bugs, especially if they are missed by testers. User stories give the developers a look into what the users are hoping to see or expecting in the future. The perspective of the user is the most important being that without the user, there is no reason for the project.

-How do I approach developing programs? What Agile processes do I hope to incorporate into my future development work?
I try to create my code in chunks. When one piece is finished, I'll test, bug fix, and test again until it feels right. Trying to knock out the whole code in one go then starting to test is a great way to brick the entire program. There would be too much to sift through to find the bugs that prevent launch. This is one reason the waterfall method fell out of style. Agile brings flexibility, as well as a continuous testing environment, that help developers fix bugs as the code is written.

-What does it mean to be a good team member in software development?
A good team member contributes what is expected while helping others. I find that people are afraid to ask questions when they don't know something, and this is a major issue in development. If something is unclear, Google it. If you can't find what you're looking for, ask someone around you. Asking questions also makes for a good team member, because if there are others with that same question, they will get the answer they need.